Move option matchers out of production code
