+++ pkg/client/kubernetes/options.go
@@ -23,21 +23,27 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
 
-type config struct {
+type Config struct {
 	clientOptions       client.Options
 	restConfig          *rest.Config
 	cacheResync         *time.Duration
 	disableCachedClient bool
 }
 
+// NewConfig returns a new Config with an empty REST config to allow testing ConfigFuncs without exporting
+// the fields of the Config type.
+func NewConfig() *Config {
+	return &Config{restConfig: &rest.Config{}}
+}
+
 // ConfigFunc is a function that mutates a Config struct.
 // It implements the functional options pattern. See
 // https://github.com/tmrts/go-patterns/blob/master/idiom/functional-options.md.
-type ConfigFunc func(config *config) error
+type ConfigFunc func(config *Config) error
 
-// WithRESTConfig returns a ConfigFunc that sets the passed rest.Config on the config object.
+// WithRESTConfig returns a ConfigFunc that sets the passed rest.Config on the Config object.
 func WithRESTConfig(restConfig *rest.Config) ConfigFunc {
-	return func(config *config) error {
+	return func(config *Config) error {
 		config.restConfig = restConfig
 		return nil
 	}
@@ -47,7 +53,7 @@ func WithRESTConfig(restConfig *rest.Config) ConfigFunc {
 // the passed ClientConnectionConfiguration.
 // The kubeconfig location in ClientConnectionConfiguration is disregarded, though!
 func WithClientConnectionOptions(cfg baseconfig.ClientConnectionConfiguration) ConfigFunc {
-	return func(config *config) error {
+	return func(config *Config) error {
 		if config.restConfig == nil {
 			return errors.New("REST config must be set before setting connection options")
 		}
@@ -59,9 +65,9 @@ func WithClientConnectionOptions(cfg baseconfig.ClientConnectionConfiguration) C
 	}
 }
 
-// WithClientOptions returns a ConfigFunc that sets the passed Options on the config object.
+// WithClientOptions returns a ConfigFunc that sets the passed Options on the Config object.
 func WithClientOptions(opt client.Options) ConfigFunc {
-	return func(config *config) error {
+	return func(config *Config) error {
 		config.clientOptions = opt
 		return nil
 	}
@@ -69,7 +75,7 @@ func WithClientOptions(opt client.Options) ConfigFunc {
 
 // WithCacheResyncPeriod returns a ConfigFunc that set the client's cache's resync period to the given duration.
 func WithCacheResyncPeriod(resync time.Duration) ConfigFunc {
-	return func(config *config) error {
+	return func(config *Config) error {
 		config.cacheResync = &resync
 		return nil
 	}
@@ -78,7 +84,7 @@ func WithCacheResyncPeriod(resync time.Duration) ConfigFunc {
 // WithDisabledCachedClient disables the cache in the controller-runtime client, so Client() will be equivalent to
 // DirectClient().
 func WithDisabledCachedClient() ConfigFunc {
-	return func(config *config) error {
+	return func(config *Config) error {
 		config.disableCachedClient = true
 		return nil
 	}
